feat: using index.js instead, to follow node package standards
