+++ package.json
@@ -2,7 +2,7 @@
   "name": "eleventy-plugin-add-web-component-definitions",
   "version": "1.1.2",
   "description": "",
-  "main": ".eleventy.js",
+  "main": "index.js",
   "keywords": [
     "eleventy",
     "11ty",
